lint the project

+++ controller/sending/sendFiles.ts
@@ -7,28 +7,27 @@ import Embedding from "@models/Users/Embedding";
 const fsPromises = fs.promises;
 const FormData = require("form-data");
 
-
 export interface ExamFileInfo {
-  examId: string;
-  chunkIndex: number;
-  lastChunk: boolean;
-  chunk?: Buffer;
-  chunkStartTime: number;
-  chunkEndTime: number;
+    examId: string;
+    chunkIndex: number;
+    lastChunk: boolean;
+    chunk?: Buffer;
+    chunkStartTime: number;
+    chunkEndTime: number;
 }
 
 export interface StorageCallback {
-  (chunk: any, examId: string, userId: string): Promise<any>;
+    (chunk: any, examId: string, userId: string): Promise<any>;
 }
 
 export interface ExamChunkResultCallback {
-  (
-    studentId: string,
-    examId: string,
-    result: string,
-    start: number,
-    end: number
-  ): Promise<any>;
+    (
+        studentId: string,
+        examId: string,
+        result: string,
+        start: number,
+        end: number
+    ): Promise<any>;
 }
 
 /**
@@ -42,52 +41,51 @@ export interface ExamChunkResultCallback {
  * @param resultCallback
  */
 export const sendExamFile = async (
-  userId: string,
-  receiverBaseUrl: string,
-  fileInfo: ExamFileInfo,
-  filePath: string,
-  embedding: Embedding,
-  resultCallback: ExamChunkResultCallback
+    userId: string,
+    receiverBaseUrl: string,
+    fileInfo: ExamFileInfo,
+    filePath: string,
+    embedding: Embedding,
+    resultCallback: ExamChunkResultCallback
 ) => {
-  if (fileInfo.chunkIndex % 2 == 0) {
-    
-    // const fileName = `${fileInfo.examId}-${userId}-${fileInfo.chunkIndex}.webm`;
+    if (fileInfo.chunkIndex % 2 == 0) {
+        // const fileName = `${fileInfo.examId}-${userId}-${fileInfo.chunkIndex}.webm`;
 
-    await sendFileHttpMethod(
-      filePath,
-      "chunk",
-      `${receiverBaseUrl}/exams/${userId}`,
-      async (res) => {
-        await resultCallback(
-          userId,
-          fileInfo.examId,
-          res,
-          fileInfo.chunkStartTime,
-          fileInfo.chunkEndTime,
+        await sendFileHttpMethod(
+            filePath,
+            "chunk",
+            `${receiverBaseUrl}/exams/${userId}`,
+            async (res) => {
+                await resultCallback(
+                    userId,
+                    fileInfo.examId,
+                    res,
+                    fileInfo.chunkStartTime,
+                    fileInfo.chunkEndTime
+                );
+            },
+            undefined,
+            embedding.vector
         );
-      },
-      undefined,
-      embedding.vector
-    );
-  }
+    }
 };
 
 export async function sendInitialVideo(
-  video: Buffer,
-  studentId: string,
-  serverBaseUrl: string,
-  resultCallback: (userId: string, embedding: number[]) => Promise<void>
+    video: Buffer,
+    studentId: string,
+    serverBaseUrl: string,
+    resultCallback: (userId: string, embedding: number[]) => Promise<void>
 ) {
-  const fileName = `video-${studentId}-${Date.now()}.webm`;
-  await sendFileHttpMethod(
-    fileName,
-    "student_video",
-    `${serverBaseUrl}/users`,
-    async (res) => {
-      await resultCallback(studentId, res.embedding);
-    },
-    video,
-  );
+    const fileName = `video-${studentId}-${Date.now()}.webm`;
+    await sendFileHttpMethod(
+        fileName,
+        "student_video",
+        `${serverBaseUrl}/users`,
+        async (res) => {
+            await resultCallback(studentId, res.embedding);
+        },
+        video
+    );
 }
 
 /**
@@ -102,43 +100,42 @@ export async function sendInitialVideo(
  * @param additionalFields
  */
 export async function sendFileHttpMethod(
-  fileName: string,
-  fieldName: string,
-  url: string,
-  callback: (res: any) => Promise<void>,
-  buffer? : Buffer,
-  additionalFields?: any,
+    fileName: string,
+    fieldName: string,
+    url: string,
+    callback: (res: any) => Promise<void>,
+    buffer?: Buffer,
+    additionalFields?: any
 ) {
-  if (buffer){
-    fileName = `/tmp/${fileName}`
-    await fsPromises.writeFile(fileName, buffer);
-  }
-  const fd = new FormData();
-  fd.append(fieldName, fs.createReadStream(fileName));
-  
-  // DELETEME
-  console.log(fs.readFileSync(fileName))
-  
-  for (const key in additionalFields) {
-    fd.append(key, additionalFields[key]);
-  }
-  try {
-    let res = await axios({
-      method: "post",
-      headers: fd.getHeaders(),
-      data: fd,
-      url: url,
-    });
-    // FIXME this will not throw error if server is closed
-    await callback(res.data);
-  } catch (e) {
-    console.error(e);
-  } finally {
-    try {
-      if (buffer)
-        await fsPromises.unlink(`${fileName}`);
-    } catch (e) {
-      console.error(e);
+    if (buffer) {
+        fileName = `/tmp/${fileName}`;
+        await fsPromises.writeFile(fileName, buffer);
     }
-  }
+    const fd = new FormData();
+    fd.append(fieldName, fs.createReadStream(fileName));
+
+    // DELETEME
+    console.log(fs.readFileSync(fileName));
+
+    for (const key in additionalFields) {
+        fd.append(key, additionalFields[key]);
+    }
+    try {
+        let res = await axios({
+            method: "post",
+            headers: fd.getHeaders(),
+            data: fd,
+            url: url,
+        });
+        // FIXME this will not throw error if server is closed
+        await callback(res.data);
+    } catch (e) {
+        console.error(e);
+    } finally {
+        try {
+            if (buffer) await fsPromises.unlink(`${fileName}`);
+        } catch (e) {
+            console.error(e);
+        }
+    }
 }
